(svn r26609) -Fix: compile warnings on MSVC for 32 bits builds due to assigning values to variables with smaller integer size

+++ src/3rdparty/squirrel/squirrel/sqfuncstate.cpp
@@ -537,7 +537,7 @@ SQFunctionProto *SQFuncState::BuildProto()
 	for(SQUnsignedInteger no = 0; no < _lineinfos.size(); no++) f->_lineinfos[no] = _lineinfos[no];
 	for(SQUnsignedInteger no = 0; no < _defaultparams.size(); no++) f->_defaultparams[no] = _defaultparams[no];
 
-	memcpy(f->_instructions,&_instructions[0],_instructions.size()*sizeof(SQInstruction));
+	memcpy(f->_instructions,&_instructions[0],(size_t)_instructions.size()*sizeof(SQInstruction));
 
 	f->_varparams = _varparams;
